Stage 2 (Scheduling enters)
Overcome: no longer need to run the file for each and every class. the Script is capable of running thoughout the week

This is Second Succesful that i created. Here aim is to Schedule All the classes so that the user don't have to run the file again and again for every class.