Stage 3 (User Input)
Overcomes: This bot will not only for me but it will works for al those who are from my class or who havve same timetable as me

This bot will take id password from the user in the terminal just for the first time and it will run for a week again
this was created keeping my mates in the view. 
like not to be so mean let's share this need with everyone in need.